requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=0.11.0=pypi_0
alabaster=0.7.12=py36_0
anyio=2.2.0=py36hecd8cb5_1
appnope=0.1.2=py36hecd8cb5_1001
argon2-cffi=20.1.0=py36h9ed2024_1
astor=0.8.1=pypi_0
astroid=2.4.2=py36_0
async_generator=1.10=py36h28b3542_0
attrs=20.3.0=pyhd3eb1b0_0
babel=2.9.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=1.5.0=pypi_0
box2d-py=2.3.8=py36h0130604_1
brotlipy=0.7.0=py36h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
ca-certificates=2020.10.14=0
certifi=2020.6.20=py36_0
cffi=1.14.4=pypi_0
chardet=4.0.0=py36hecd8cb5_1003
click=7.1.2=pypi_0
cloudpickle=1.6.0=py_0
contextvars=2.4=py_0
cryptography=3.4.7=py36h2fd3fbb_0
cycler=0.10.0=pypi_0
dataclasses=0.7=py36_0
dbus=1.13.18=h18a8e69_0
decorator=4.4.2=pypi_0
defusedxml=0.6.0=pypi_0
docopt=0.6.2=pypi_0
docutils=0.16=py36_1
entrypoints=0.3=pypi_0
expat=2.2.10=hb1e8313_2
ffmpeg=4.2.2=h97e5cf8_0
flask=1.1.2=pypi_0
freetype=2.10.4=ha233b18_0
future=0.18.2=py36_1
gast=0.4.0=pypi_0
gettext=0.19.8.1=hb0f4f8b_2
glib=2.66.1=h9bbe63b_0
gmp=6.2.1=h23ab428_2
gnutls=3.6.5=h91ad68e_1002
grip=4.5.2=pypi_0
grpcio=1.11.0=pypi_0
gym=0.18.0=py36h9cf137f_0
html5lib=0.9999999=pypi_0
icu=58.2=h0a44026_3
idna=2.10=pyhd3eb1b0_0
imagesize=1.2.0=py_0
immutables=0.15=py36h9ed2024_0
importlib-metadata=3.4.0=pypi_0
importlib_metadata=3.10.0=hd3eb1b0_0
iniconfig=1.1.1=pypi_0
intel-openmp=2019.4=233
ipykernel=5.4.3=pypi_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pypi_0
isort=5.6.4=py_0
itsdangerous=1.1.0=pypi_0
jedi=0.18.0=pypi_0
jinja2=2.11.3=pyhd3eb1b0_0
jpeg=9b=he5867d9_2
jsanimation=0.1=pypi_0
json5=0.9.5=py_0
jsonschema=3.2.0=py_2
jupyter=1.0.0=pypi_0
jupyter-client=6.1.11=pypi_0
jupyter-console=6.2.0=pypi_0
jupyter-packaging=0.7.12=pyhd3eb1b0_0
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py36hecd8cb5_0
jupyter_server=1.4.1=py36hecd8cb5_0
jupyterlab=3.0.11=pyhd3eb1b0_1
jupyterlab-widgets=1.0.0=pypi_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.4.0=pyhd3eb1b0_0
keyring=21.4.0=py36_1
kiwisolver=1.3.1=py36h23ab428_0
lame=3.100=h1de35cc_0
lazy-object-proxy=1.4.3=py36h1de35cc_0
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.1.0=hcb84e12_1
libuv=1.40.0=haf1e3a3_0
libvpx=1.7.0=h378b8a2_0
lz4-c=1.9.3=h23ab428_0
markdown=3.3.3=pypi_0
markupsafe=1.1.1=py36h1de35cc_0
matplotlib=3.3.4=py36hecd8cb5_0
matplotlib-base=3.3.4=py36h8b3ea08_0
mccabe=0.6.1=py36_1
mistune=0.8.4=py36h1de35cc_0
mkl=2019.4=233
mkl-service=2.3.0=py36h9ed2024_0
mkl_fft=1.2.1=py36ha059aab_0
mkl_random=1.1.1=py36h959d312_0
nbclassic=0.2.6=pyhd3eb1b0_0
nbclient=0.5.1=pypi_0
nbconvert=6.0.7=py36_0
nbformat=5.1.2=pypi_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.4.1=h3018a27_0
ninja=1.10.2=py36hf7b0b51_0
notebook=6.2.0=pypi_0
numpy=1.19.2=py36h456fd55_0
numpy-base=1.19.2=py36hcfb5961_0
numpydoc=1.1.0=py_0
olefile=0.46=py36_0
openh264=2.1.0=hd9629dc_0
openssl=1.1.1k=h9ed2024_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.1.5=pypi_0
pandoc=2.12=hecd8cb5_0
pandocfilters=1.4.3=py36hecd8cb5_1
parso=0.8.1=pypi_0
path-and-address=2.0.1=pypi_0
pcre=8.44=hb1e8313_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.0=py36h5270095_0
pip=21.0.1=py36hecd8cb5_0
pluggy=0.13.1=pypi_0
prometheus-client=0.9.0=pypi_0
prometheus_client=0.10.0=pyhd3eb1b0_0
prompt-toolkit=3.0.14=pypi_0
protobuf=3.5.2=pypi_0
psutil=5.7.2=py36haf1e3a3_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py=1.10.0=pypi_0
pycodestyle=2.6.0=py_0
pycparser=2.20=py_2
pyflakes=2.2.0=py_0
pyglet=1.5.14=py36h79c6626_1
pygments=2.7.4=pypi_0
pylint=2.6.0=py36_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py36h655552a_0
pyrsistent=0.17.3=py36haf1e3a3_0
pysocks=1.7.1=py36hecd8cb5_0
pytest=6.2.2=pypi_0
python=3.6.12=h26836e1_2
python-dateutil=2.8.1=pyhd3eb1b0_0
python.app=2=py36_10
python_abi=3.6=1_cp36m
pytorch=1.7.1=py3.6_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=pypi_0
pyzmq=22.0.2=pypi_0
qt=5.9.7=h468cd18_1
qtawesome=1.0.1=py_0
qtconsole=5.0.2=pypi_0
qtpy=1.9.0=py_0
readline=8.1=h9ed2024_0
requests=2.25.1=pyhd3eb1b0_0
rope=0.18.0=py_0
scipy=1.5.4=pypi_0
seaborn=0.11.1=pyhd3eb1b0_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py36hecd8cb5_0
sip=4.19.24=py36hb1e8313_0
six=1.15.0=py36hecd8cb5_0
sniffio=1.2.0=py36hecd8cb5_1
snowballstemmer=2.0.0=py_0
sphinx=3.2.1=py_0
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=1.0.3=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.4=py_0
spyder=3.3.6=py36_0
spyder-kernels=0.5.2=py36_0
sqlite=3.33.0=hffcf06c_0
tensorboard=1.7.0=pypi_0
tensorflow=1.7.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.2=pypi_0
testpath=0.4.4=pyhd3eb1b0_0
tk=8.6.10=hb0a8c7a_0
toml=0.10.2=pypi_0
torchaudio=0.7.2=py36
torchvision=0.8.2=py36_cpu
tornado=6.1=py36h9ed2024_0
traitlets=4.3.3=py36_0
typed-ast=1.4.1=py36h1de35cc_0
typing_extensions=3.7.4.3=pyha847dfd_0
unityagents=0.4.0=pypi_0
urllib3=1.26.4=pyhd3eb1b0_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py36_1
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
wrapt=1.11.2=py36h1de35cc_0
wurlitzer=2.0.1=py36_0
x264=1!157.20191217=h1de35cc_0
xz=5.2.5=h1de35cc_0
zeromq=4.3.4=h23ab428_0
zipp=3.4.0=pypi_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
